requirements.txt
Django==4.1
PyJWT==2.4.0
requests
